@@ -4,7 +4,7 @@ version: 1.0.0
 homepage: https://github.com/lijungegea/flutter_simple_image_crop.git
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -16,8 +16,9 @@ dev_dependencies:
 
 flutter:
   plugin:
-    android:
-      package: com.lykhonis.simpleimagecrop
-      pluginClass: SimpleImageCropPlugin
-    ios:
-      pluginClass: SimpleImageCropPlugin
+    platforms:
+      android:
+        package: com.lykhonis.simpleimagecrop
+        pluginClass: SimpleImageCropPlugin
+      ios:
+        pluginClass: SimpleImageCropPlugin
